requirements.txt
flask
pyodbc
sqlalchemy
flask-bootstrap
flask-wtf
flask-testing
blinker
flask-login
python-dotenv